feat(client) fix pending changes to show for each service
