+++ packages/amplication-client/src/VersionControl/PendingChangesPage.scss
@@ -11,6 +11,11 @@
     }
   }
 
+  &__title {
+    @include body3;
+    margin-bottom: var(--large-spacing);
+  }
+
   .pending-change-with-compare {
     margin-left: 0;
     margin-right: 0;
